refactor: make the EventDispatcher an interface and create a DefaultEventDispatcher class, allow clients to change the eventDispatcher via the GatewayGrpcClient api
